Feature/#110 권한 설정이 필요한 모든 코드를 수정한다 (#123)
* feat : 팀 생성 권한 추가

* feat : 팀 업데이트 권한 설정
- 팀장만 팀 정보를 업데이트 할 수 있습니다.

* test : 팀 정보 수정 권한 확인 테스트

* feat : 팀 이미지 업데이트 & 팀 삭제 권한 적용

* feat : 팀 join 코드 todo 내용 추가
- 초대된 member는 ROLE_팀원 직위가 부여됩니다.
- 이미 가입된 팀원은 가입할 수 없습니다.

* feat : 나의 팀 조회에 member 검증 추가

* fix : 포맷팅 누락된 코드 수정

* feat : 다른 사람의 팀 조회는 불가능하도록 변경

* fix: 메서드의 의미를 담은 이름으로 메서드명 변경

* fix : 스터디 생성 권한 수정 및 스터디장 권한 부여 및 오타 수정

* feat: 스터디 삭제 권한 수정
- 스터디 삭제에 대한 권한을 수정하였습니다.
- `나의 스터디 조회` 메서드명을 `나의 팀 조회` 네이밍 로직과 동일하게 변경하였습니다.
- 테스트에 불필요한 코드를 삭제하였습니다.

* fix: 스터디 조회 잘못 수정된 부분 재수정

* fix : 스터디 수정 권한 수정

* fix: 스터디 조회 관련 재확인이 필요하여 원래 코드로 되돌림

* feat: 스터디 권한 추가

* test: 스터디 테스트 코드 수정
- setup에 스터디를 추가하며 스터디 repository에 저장된 스터디 개수가 늘어남에 따라 테스트 코드를 수정함

* fix : 스터디 권한 마무리 수정

* feat : 커리큘럼 권한 부여

* test : 커리큘럼 권한 부여에 따른 테스트 수정

* feat: 참여자 api 권한 부여

* test: 참여자 권한 부여에 따른 테스트 수정

* test: 수정될 코드 임시 disable 처리

* feat: WebMvcConfig 추가

* fix : LoginMember annotation이 적용되지 않던 문제 해결

* test: mock용 테스트 데이터 추가

* test: 권한 수정에 따른 token이 필요한 test 수정 - Study part

* test : 테스트에 토큰 추가

* test : controller test에 token 추가

* fix : document 권한 처리 최종

* chore : 개행 및 기타 수정

* fix : 스터디 권한 최종 수정

* fix : 커리큘럼 권한 수정 최종

* fix : 참여자 권한 수정 최종

* fix : 팀 관련 권한 수정 최종

* chore : member 권한 주석 추가

* chore : MemberTeam은 다른 pr 로 진행할 예정이므로 수정했던 사항 재수정

* chore : 접근 제어자 및 개행 수정

* fix : 참여자 탈퇴 기능 권한 수정
- 참여자 틸퇴 기능 권한이 스터디장에서 스터디원으로 수정됩니다.

* chore : StudyMember -> participant 수정

* fix : 학습자료 수정과 삭제는 해당 자료 업로더만 가능하도록 권한 수정

* feat : findMyTeamsAndStudies method 권한 부여

+++ src/main/java/doore/document/application/DocumentCommandService.java
@@ -1,7 +1,12 @@
 package doore.document.application;
 
-import static doore.document.domain.DocumentGroupType.*;
-import static doore.document.exception.DocumentExceptionType.*;
+import static doore.document.domain.DocumentGroupType.STUDY;
+import static doore.document.domain.DocumentGroupType.TEAM;
+import static doore.document.exception.DocumentExceptionType.INVALID_DOCUMENT_TYPE;
+import static doore.document.exception.DocumentExceptionType.LINK_DOCUMENT_NEEDS_URL;
+import static doore.document.exception.DocumentExceptionType.NOT_FOUND_DOCUMENT;
+import static doore.document.exception.DocumentExceptionType.NO_FILE_ATTACHED;
+import static doore.member.exception.MemberExceptionType.UNAUTHORIZED;
 import static doore.study.exception.StudyExceptionType.NOT_FOUND_STUDY;
 import static doore.team.exception.TeamExceptionType.NOT_FOUND_TEAM;
 
@@ -16,6 +21,9 @@ import doore.document.domain.repository.FileRepository;
 import doore.document.exception.DocumentException;
 import doore.file.application.S3DocumentFileService;
 import doore.file.application.S3ImageFileService;
+import doore.member.domain.repository.MemberRepository;
+import doore.member.exception.MemberException;
+
 import doore.garden.domain.Garden;
 import doore.garden.domain.GardenType;
 import doore.garden.domain.repository.GardenRepository;
@@ -38,12 +46,14 @@ public class DocumentCommandService {
     private final TeamRepository teamRepository;
     private final StudyRepository studyRepository;
     private final FileRepository fileRepository;
+    private final MemberRepository memberRepository;
     private final S3ImageFileService s3ImageFileService;
     private final S3DocumentFileService s3DocumentFileService;
     private final GardenRepository gardenRepository;
 
     public void createDocument(DocumentCreateRequest request, List<MultipartFile> multipartFiles,
-                               DocumentGroupType groupType, Long groupId) {
+                               DocumentGroupType groupType, Long groupId, Long memberId) {
+        validateExistMember(memberId);
         validateExistGroup(groupType, groupId);
         validateDocumentType(request.type(), request.url(), multipartFiles);
         Document document = Document.from(request, groupType, groupId);
@@ -123,13 +133,21 @@ public class DocumentCommandService {
         gardenRepository.save(garden);
     }
 
-    public void updateDocument(DocumentUpdateRequest request, Long documentId) {
+    public void updateDocument(DocumentUpdateRequest request, Long documentId, Long memberId) {
+        validateExistMember(memberId);
         Document document = validateExistDocument(documentId);
+        if (!document.isMine(memberId)){
+            throw new MemberException(UNAUTHORIZED);
+        }
         document.update(request.title(), request.description(), request.accessType());
     }
 
-    public void deleteDocument(Long documentId) {
+    public void deleteDocument(Long documentId, Long memberId) {
+        validateExistMember(memberId);
         Document document = validateExistDocument(documentId);
+        if (!document.isMine(memberId)){
+            throw new MemberException(UNAUTHORIZED);
+        }
         deleteGarden(document);
         documentRepository.deleteById(documentId);
     }
@@ -142,5 +160,9 @@ public class DocumentCommandService {
 
     private Document validateExistDocument(Long documentId) {
         return documentRepository.findById(documentId).orElseThrow(() -> new DocumentException(NOT_FOUND_DOCUMENT));
+    }
+
+    private void validateExistMember(Long memberId) {
+        memberRepository.findById(memberId).orElseThrow(() -> new MemberException(UNAUTHORIZED));
     }
 }
